2003-07-14 21:15 UTC+0200 Giancarlo Niccolai <gian@niccolai.ws>   * include/thread.h   * source/vm/codebloc.c   * source/vm/memvars.c   * source/vm/thread.c   * source/vm/itemapi.c   * include/thread.h     ! thread safed private and memvars     + Added hb_itemClearMT() function for some that is able       to get another thread's stack to clear a Memvar. This requires       cooperation of thread.c and memvars.c (that have now some MT specific       functions).
  * source/vm/estack.c
    ! exit sequence uses hb_threadDestroyStack() (easier developement)

  * source/vm/extend.c
    ! Fixed a warning on longlong

  * source/vm/dynsym.c
    ! Put dynsym threadsafe (at least scanning for existing dynsym)
    % Optimized HB_VM_STACK access. (it FLIES :-)

  * source/vm/inet.c
    ! re-added the signal handler for linux, but using a signal
      that won't be used later on by the service signal handlers

  * source/pp/ppcore.c
    ! Fixed warning in preprocessor

  + tests/mtstress.prg
    ! Stress tests for new MT features.

+++ xharbour/source/vm/extend.c
@@ -1,5 +1,5 @@
 /*
- * $Id: extend.c,v 1.23 2003/07/07 03:30:49 lculik Exp $
+ * $Id: extend.c,v 1.24 2003/07/13 18:15:40 walito Exp $
  */
 
 /*
@@ -595,7 +595,7 @@ void *hb_parptr( int iParam, ... )
 #ifndef HB_LONG_LONG_OFF
       else if( HB_IS_LONGLONG( pItem ) )
       {
-         return ( void * ) pItem->item.asLongLong.value;
+         return ( void * ) ((long)pItem->item.asLongLong.value);
       }
 #endif
       else if( HB_IS_ARRAY( pItem ) )
